+++ Assets/_Scripts/Player/Damage/Weapon.cs
@@ -9,32 +9,27 @@ namespace _Scripts.Player
     {
         [SerializeField] private GameObject player;
         [SerializeField] private GameObject bombPrefab;
+        [SerializeField] private float fireRate;
 
         private int _damage;
+        private float _nextTimeToFire = 0f;
 
         public Camera fpsCam;
         public LayerMask layerMask;
 
-        // public int damage = 10;
         public float range = 100f;
 
         public float throwForce;
 
-        //void Start()
-        //{
-        //    _damage = player.GetComponent<PlayerController>().Stats.Damage;
-        //}
-
-        // Update is called once per frame
         void Update()
         {
             // Get the damage the player can make
             _damage = player.GetComponent<Stats>().Damage;
             
             // If Fire1 is pressen and player is in Default Mode
-            if (Input.GetButtonDown("Fire1") && player.GetComponent<PlayerController>().GetCurrentPlayerMode() == 0)
+            if (Input.GetButton("Fire1") && Time.time >= _nextTimeToFire && player.GetComponent<PlayerController>().GetPlayerMode() == PlayerController.PlayerMode.Default)
             {
-                Debug.Log("Shoot");
+                _nextTimeToFire = Time.time + 1f / fireRate;
                 Shoot();
             }
 
@@ -54,6 +49,7 @@ namespace _Scripts.Player
                 // If a building is hit
                 if (hit.transform.gameObject.CompareTag("Buildable"))
                 {
+                    Debug.Log("Hit Buildable");
                     hit.transform.gameObject.GetComponent<Buildable>().TakeDamage(_damage);
                 }